+++ Source/Core/Macros.hpp
@@ -154,6 +154,7 @@
 
 
 
+
 /// Logs current function for debugging
 #if defined(ENIGMA_DEBUG)
 //#define ENIGMA_TRACE_CURRENT_FUNCTION() ENIGMA_TRACE(::Enigma::StringUtils::Cleanup(::Enigma::StringUtils::Cleanup(::Enigma::StringUtils::Cleanup(ENIGMA_CURRENT_FUNCTION, " __cdecl"), "struct"), "class"))
@@ -163,6 +164,22 @@
 #endif
 ///
 
+
+
+
+//void TimedFunction(
+//	const std::function<void()>& onStart,
+//	const std::function<void()>& func,
+//	const std::function<void(double)>& onFinish)
+//{
+//	onStart();
+//	ENIGMA_BEGIN_TIMER(t1);
+//		func();
+//	auto elapsed = ENIGMA_END_TIMER(t1, double, std::milli) / 1000.0;
+//	onFinish(elapsed);
+//}
+//
+//
 
 
 #endif // !ENIGMA_MACROS_H