+++ Source/Scenes/EncryptFileScene.cpp
@@ -160,7 +160,7 @@ void EncryptFileScene::OnImGuiDraw()
 			if (m_save_to_database)
 			{
 				ImGui::Text("Encryption Title:");
-				ImGuiWidgets::InputTextWithHint("##idb", "(e.g: My Image) helps with searching through encryption records in the future", &m_db_title, win_w / 1.3f);
+				ImGuiWidgets::InputTextWithHint("##idb", "(e.g: An important video) helps with searching through encryption records in the future", &m_db_title, win_w / 1.3f);
 			}
 		}
 		ImGui::PopFont();